refactor: inline settings/onboarding/home sub-files
